refactor: Extract MockOIDCProvider to separate utility file to reduce test file size
